chore: 프로젝트 디렉토리 구조, 이름 변경 (#7)
- github 기능 이용하여 프론트와 백으로 나누면서 깊어진 디렉토리 구조 단순하게 변경
- Spring 프로젝트 이름 demo 에서 backend 로 변경
- .gitignore에 .DS_Store 추가
