--- backend/demo/src/main/java/com/snacks/demo/DemoApplication.java
@@ -1,13 +1,0 @@
-package com.snacks.demo;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class DemoApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(DemoApplication.class, args);
-	}
-
-}
