--- backend/demo/src/main/java/com/snacks/demo/controller/UserController.java
@@ -1,25 +1,0 @@
-package com.snacks.demo.controller;
-
-import com.snacks.demo.dto.UserDto;
-import com.snacks.demo.response.ResponseService;
-import com.snacks.demo.service.AuthService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.validation.BindingResult;
-import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/users")
-public class UserController {
-
-  @GetMapping("/test")
-  public String test(){
-    return "/users/test fin";
-  }
-}
