--- backend/demo/src/main/java/com/snacks/demo/entity/User.java
@@ -1,28 +1,0 @@
-package com.snacks.demo.entity;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.validation.constraints.NotNull;
-import lombok.Getter;
-import lombok.Setter;
-
-@Entity
-@Getter
-@Setter
-public class User {
-  @Id
-  @Column(name = "id")
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long id;
-
-  @Column(name = "email")
-  @NotNull
-  private String email;
-
-  @Column(name = "password")
-  @NotNull
-  private String password;
-}
