--- backend/demo/src/main/java/com/snacks/demo/jwt/auth/CustomUserDetails.java
@@ -1,56 +1,0 @@
-package com.snacks.demo.jwt.auth;
-
-import com.snacks.demo.entity.User;
-import java.util.Collection;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-public class CustomUserDetails implements UserDetails {
-
-  @Autowired
-  private User user;
-
-  public CustomUserDetails(User user) {
-    this.user = user;
-  }
-
-  public User getUser() {
-    return user;
-  }
-
-  @Override
-  public Collection<? extends GrantedAuthority> getAuthorities() {
-    return null;
-  }
-
-  @Override
-  public String getPassword() {
-    return user.getPassword();
-  }
-
-  @Override
-  public String getUsername() {
-    return user.getEmail();
-  }
-
-  @Override
-  public boolean isAccountNonExpired() {
-    return true;
-  }
-
-  @Override
-  public boolean isAccountNonLocked() {
-    return true;
-  }
-
-  @Override
-  public boolean isCredentialsNonExpired() {
-    return true;
-  }
-
-  @Override
-  public boolean isEnabled() {
-    return true;
-  }
-}
