--- backend/demo/src/main/java/com/snacks/demo/response/CommonResponse.java
@@ -1,7 +1,0 @@
-package com.snacks.demo.response;
-
-public class CommonResponse {
-    public String log;
-}
-
-
