--- backend/demo/src/main/java/com/snacks/demo/response/ListResponse.java
@@ -1,5 +1,0 @@
-package com.snacks.demo.response;
-
-public class ListResponse<T> extends CommonResponse {
-   public  T dataList;
-}
