--- backend/demo/src/main/java/com/snacks/demo/response/SingleResponse.java
@@ -1,5 +1,0 @@
-package com.snacks.demo.response;
-
-public class SingleResponse<T> extends CommonResponse {
-    T data;
-}
