--- backend/demo/src/main/java/com/snacks/demo/response/TokenResponse.java
@@ -1,7 +1,0 @@
-package com.snacks.demo.response;
-
-public class TokenResponse {
-
-  public String refresh_token;
-  public String access_token;
-}
