--- backend/demo/src/test/java/com/snacks/demo/ConstantTest.java
@@ -1,7 +1,0 @@
-package com.snacks.demo;
-
-public class ConstantTest {
-
-  public static final String VALID_EMAIL = "valid@email.com";
-  public static final String VALID_PASSWORD = "1234";
-}
