+++ src/main/java/com/snacks/demo/controller/UserController.java
@@ -1,0 +1,25 @@
+package com.snacks.backend.controller;
+
+import com.snacks.backend.dto.UserDto;
+import com.snacks.backend.response.ResponseService;
+import com.snacks.backend.service.AuthService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.validation.BindingResult;
+import org.springframework.validation.annotation.Validated;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/users")
+public class UserController {
+
+  @GetMapping("/test")
+  public String test(){
+    return "/users/test fin";
+  }
+}
