+++ src/main/java/com/snacks/demo/repository/AuthRepository.java
@@ -1,0 +1,14 @@
+package com.snacks.backend.repository;
+
+import com.snacks.backend.entity.User;
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface AuthRepository extends JpaRepository<User, Long> {
+  Optional<User> findByEmail(String email);
+
+  User save(User user);
+
+}
