+++ src/main/java/com/snacks/demo/response/CommonResponse.java
@@ -1,0 +1,7 @@
+package com.snacks.backend.response;
+
+public class CommonResponse {
+    public String log;
+}
+
+
