+++ src/main/java/com/snacks/demo/response/ListResponse.java
@@ -1,0 +1,5 @@
+package com.snacks.backend.response;
+
+public class ListResponse<T> extends CommonResponse {
+   public  T dataList;
+}
