+++ src/main/java/com/snacks/demo/response/SingleResponse.java
@@ -1,0 +1,5 @@
+package com.snacks.backend.response;
+
+public class SingleResponse<T> extends CommonResponse {
+    T data;
+}
