+++ src/main/java/com/snacks/demo/response/TokenResponse.java
@@ -1,0 +1,7 @@
+package com.snacks.backend.response;
+
+public class TokenResponse {
+
+  public String refresh_token;
+  public String access_token;
+}
