+++ src/test/java/com/snacks/demo/BackendApplicationTests.java
@@ -1,0 +1,13 @@
+package com.snacks.backend;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class BackendApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
