+++ src/test/java/com/snacks/demo/ConstantTest.java
@@ -1,0 +1,7 @@
+package com.snacks.backend;
+
+public class ConstantTest {
+
+  public static final String VALID_EMAIL = "valid@email.com";
+  public static final String VALID_PASSWORD = "1234";
+}
